xyz coord to voxel completed
Implementation of converting a set of xyz coordinates to a voxel matrix has been completed, a user interface added to demonstrate how this tool can be used and some bugs corrected. Draft user guide updated.

+++ Source/Modules/Module_Voxel_Byte/Voxel_hcp_object/parameters.h
@@ -9,7 +9,8 @@
 enum class data_storage_type_enum { red, green, blue, alpha, value };
 
 //typedef uint8_t voxel_data_type;
-typedef uint8_t voxel_element_data_type; // For color r,g,b, and alpha or voxel value data type
+//typedef uint8_t voxel_element_data_type; // For color r,g,b, and alpha or voxel value data type
+typedef int voxel_element_data_type; // For color r,g,b, and alpha or voxel value data type
 
 // voxel_data_type and voxel_GLdata_type must be of the same equivalent data type 
 typedef int   voxel_data_type;   // CPU
